Big Update

+++ routes/grouptasks.js
@@ -1,14 +1,14 @@
-const validateObjectId = require('../middleware/validateObjectId');
 const auth = require('../middleware/auth');
 const admin = require('../middleware/group-admin');
 const {Task, validate} = require('../models/task');
 const mongoose = require('mongoose');
 const express = require('express');
 const router = express.Router();
+const {Group} = require('../models/group');
 
-router.get('/', auth, async (req, res) => {
-  const task = await Task.find({ owner: req.user._id});
-  res.send(task);
+router.get('/:id', auth, async (req, res) => {
+  const grouptasks = await Task.find({ owner: req.params.id});
+  res.send(grouptasks);
 });
 
 router.post('/', auth, async (req, res) => {
@@ -20,36 +20,47 @@ router.post('/', auth, async (req, res) => {
     description: req.body.description,
     status: req.body.status || "to-do",
     tags: req.body.tags,
-    owner: req.user._id
+    owner: req.body.owner
   });
-  if (!task.tags) return res.status(400).send('Task should have and least one tag');
+  const check = await Group
+    .findById(task.owner);
+
+  if (req.user._id !== check.groupAdmin) return res.status(403).send('Access denied.');
+
   task = await task.save();
   
   res.send(task);
 });
 
-router.put('/:id',[auth,validateObjectId], async (req, res) => {
+router.put('/:id', auth, async (req, res) => {
   const { error } = validate(req.body); 
   if (error) return res.status(400).send(error.details[0].message);
+  const taskToCheck = await Task
+    .findById(req.params.id);
 
-  const old = await Task.findById(req.params.id);
-  console.log(old);
-  
-  const task = await Task.findByIdAndUpdate(req.params.id, {
+  const group = await Group
+    .findById (taskToCheck.owner);
     
+  if (!group.members.includes(req.user._id)) res.status(403).send('Access denied.');
+
+  const task = await Task.findByIdAndUpdate(req.params.id, {
         title: req.body.title,
         description: req.body.description,
         status: req.body.status,
-        tag: req.body.tag,
-      },
-      {  new: true,}
-      );
+        tags: req.body.tags,
+      });
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
+  
   res.send(task);
 });
 
-router.delete('/:id', [auth, validateObjectId], async (req, res) => {
+router.delete('/:id', auth, async (req, res) => {
+  const check = await Group
+    .findById(req.params.id);
+
+  if (req.user._id !== check.groupAdmin) return res.status(403).send('Access denied.');
+
   const task = await Task.findByIdAndRemove(req.params.id);
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
@@ -57,7 +68,7 @@ router.delete('/:id', [auth, validateObjectId], async (req, res) => {
   res.send(task);
 });
 
-router.get('/:id', async (req, res) => {
+router.get('/task/:id', auth, async (req, res) => {
   const task = await Task.findById(req.params.id);
 
   if (!task) return res.status(404).send('The task with the given ID was not found.');
